app/urls.py
from django.urls.conf import include, path
from rest_framework import routers, serializers, viewsets

from .api import CategoryViewSet

router = routers.DefaultRouter()
router.register(r'categories', CategoryViewSet, basename='category')

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = router.urls
