pyproject.toml
[project]
authors = [
  {name = "", email = ""},
]
dependencies = [
    "asgiref==3.4.1; python_version >= \"3.6\"",
    "django==3.2.7; python_version >= \"3.6\"",
    "pytz==2021.1",
    "sqlparse==0.4.2; python_version >= \"3.5\"",
    "autopep8>=1.6.0",
    "djangorestframework>=3.13.1",
    "markdown>=3.3.6",
]
description = ""
license = {text = "MIT"}
name = ""
requires-python = ">=3.9"
version = ""

[project.urls]
homepage = ""

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517"]

[tool]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true
